requirements.txt
django
python-dotenv
psycopg2
django-debug-toolbar
djangorestframework
djoser
drf-yasg